requirements.txt
requests>=2.28.1
beautifulsoup4>=4.11.1
pandas>=1.4.2
tqdm>=4.64.0
python-dotenv>=0.20.0
selenium>=4.10.0
pyperclip>=1.8.2
webdriver-manager>=3.8.5
gspread>=5.8.0
oauth2client>=4.1.2
aiohttp>=3.8.1
nest_asyncio>=1.5.5
youtube-transcript-api==0.6.2
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
youtube-transcript-api>=0.6.0